--- src/_hook/use-user.js
@@ -1,14 +1,0 @@
-import { useContext } from "react";
-import { UserActionContext, UserContext } from "/src/_provider/UserProvider";
-
-export const useUser = () => {
-  const user = useContext(UserContext);
-
-  return user;
-};
-
-export const useUserAction = () => {
-  const userAction = useContext(UserActionContext);
-
-  return userAction;
-};
