--- src/_query/user/use-user-query.js
@@ -1,16 +1,0 @@
-import { useEffect } from "react";
-import { useUserAction } from "/src/_hook/use-user";
-import { getUser } from "/src/_api/user/api";
-
-export default function useUserQuery() {
-  const { setUser } = useUserAction();
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      const user = await getUser();
-      setUser(user);
-    };
-
-    fetchUser();
-  }, []);
-}
