--- src/home/_component/ServiceIntroCard.jsx
@@ -1,24 +1,0 @@
-import Card from "/src/_component/common/Card";
-
-export default function ServiceIntroCard({ title, description, thumbnail }) {
-  return (
-    <Card className="tablet:service-intro flex h-max flex-col gap-[2rem] px-[3.2rem] py-[4rem]">
-      <Card.Title
-        title={{
-          headType: "h2",
-          content: title,
-        }}
-        className="tablet:service-intro-title order-1 text-[2.4rem] font-bold leading-[1.3] tracking-[-0.03rem]"
-      />
-      <Card.Description
-        description={description}
-        className="tablet:service-intro-description order-3 text-[1.6rem] font-medium leading-[150%] text-[#6b6b6b]"
-      />
-      <Card.Thumbnail
-        src={thumbnail}
-        className="tablet:service-intro-content-image order-2 h-auto w-full"
-        loading="lazy"
-      />
-    </Card>
-  );
-}
