--- src/shared/_component/LinkInfoCardList.jsx
@@ -1,15 +1,0 @@
-import LinkInfoCard from "./LinkInfoCard";
-
-export default function LinkInfoCardList({ linkInfoList }) {
-  return (
-    <ul className="grid-cols-fill-max-34 relative grid w-full justify-center gap-x-[2rem] gap-y-[2.5rem]">
-      {linkInfoList.map((linkInfo) => (
-        <li key={linkInfo.id}>
-          <a href={linkInfo.url} target="_blank">
-            <LinkInfoCard linkInfo={linkInfo} />
-          </a>
-        </li>
-      ))}
-    </ul>
-  );
-}
